🔥 (optimize-packing) invisible casting

+++ test-foundry/TestStessHeapOrdering.t.sol
@@ -16,14 +16,14 @@ contract HeapStorage {
             address id = address(uint160(i + 1));
             heap.accounts.push(HeapOrdering.Account(id, uint96(TESTED_SIZE - i)));
             heap.ranks[id] = heap.accounts.length;
-            heap.size = uint96(MAX_SORTED_USERS);
+            heap.size = MAX_SORTED_USERS;
         }
     }
 
     function update(
         address _id,
-        uint96 _formerValue,
-        uint96 _newValue
+        uint256 _formerValue,
+        uint256 _newValue
     ) public {
         HeapOrdering.update(heap, _id, _formerValue, _newValue, MAX_SORTED_USERS);
     }
@@ -31,13 +31,13 @@ contract HeapStorage {
 
 contract TestStressHeapOrdering is DSTest {
     HeapStorage public hs = new HeapStorage();
-    uint96 public ts;
-    uint96 public im;
+    uint256 public ts;
+    uint256 public im;
 
     function setUp() public {
         hs.setUp();
-        ts = uint96(hs.TESTED_SIZE());
-        im = uint96(hs.incrementAmount());
+        ts = hs.TESTED_SIZE();
+        im = hs.incrementAmount();
     }
 
     function testInsertOneTop() public {
@@ -57,12 +57,12 @@ contract TestStressHeapOrdering is DSTest {
     }
 
     function testRemoveOneMiddle() public {
-        uint96 middle = ts / 2;
+        uint256 middle = ts / 2;
         hs.update(address(uint160(middle + 1)), ts - middle, 0);
     }
 
     function testRemoveOneBottom() public {
-        uint96 end = ts - 2 * im;
+        uint256 end = ts - 2 * im;
         hs.update(address(uint160(end + 1)), ts - end, 0);
     }
 
@@ -71,12 +71,12 @@ contract TestStressHeapOrdering is DSTest {
     }
 
     function testIncreaseOneMiddle() public {
-        uint96 middle = ts / 2;
+        uint256 middle = ts / 2;
         hs.update(address(uint160(middle + 1)), ts - middle, ts - middle + im);
     }
 
     function testIncreaseOneBottom() public {
-        uint96 end = ts - 2 * im;
+        uint256 end = ts - 2 * im;
         hs.update(address(uint160(end + 1)), ts - end, ts - end + im);
     }
 
@@ -85,12 +85,12 @@ contract TestStressHeapOrdering is DSTest {
     }
 
     function testDecreaseOneMiddle() public {
-        uint96 middle = ts / 2;
+        uint256 middle = ts / 2;
         hs.update(address(uint160(middle + 1)), ts - middle, 1);
     }
 
     function testDecreaseOneBottom() public {
-        uint96 end = ts - 2 * im;
+        uint256 end = ts - 2 * im;
         hs.update(address(uint160(end + 1)), ts - end, ts - end - im);
     }
 }
